Stop specifying `default` module for tests.
To prevent themes (such as our material theme module) from affecting
this MFA module's tests, we have these tests use the default theme.
However, we had been doing so with `THEME_USE=default:default`, but
that also specifies a module of `default`. All we needed to specify was a
theme of `default`.