Install any new dependencies in local IdP before running tests
If any new composer dependencies were added to this project, the tests
will fail because the copies of ssp-base used to run the tests won't have
those new dependencies.

This update gets composer to install the latest develop version of this repo,
including its dependencies, which will at least get the tests to pass once
these updates make it into the develop branch (such as when doing a
release).